Moved unittest.c.
